Debug: Refactoring. #292
